+++ Makefile
@@ -1,10 +1,26 @@
-obj-m := src/main.o
+# Module name
+ROOTKIT		    := rootkit
+
+# Linux kernel files
+MODULEDIR	    := /lib/modules/$(shell uname -r)
+BUILDDIR	    := $(MODULEDIR)/build
+
+# Module Headers
+# HEADERS	        := src/headers
+# ccflags-y	    := -I$(HEADERS)
+
+# Module Sources
+$(ROOTKIT)-y 	+= src/main.o
+$(ROOTKIT)-y 	+= src/networking.o
+
+# Module output
+obj-m           := $(ROOTKIT).o
 
 all:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
+	$(MAKE) -C $(BUILDDIR) M=$(PWD) modules
 
 clean:
-	make -C /lib/modules/$(shell uname -r)/build M=$(pwd) clean
+	$(MAKE) -C $(BUILDDIR) M=$(PWD) clean
 
 watch:
 	sudo dmesg -C
@@ -12,5 +28,5 @@ watch:
 
 test:
 	sudo dmesg -C
-	sudo insmod src/main.ko
+	sudo insmod ${ROOTKIT}.ko
 	dmesg -w
