+++ src/headers/networking.h
@@ -1,0 +1,6 @@
+#ifndef NETWORKING_H
+#define NETWORKING_H
+
+int MRK_init_nethook(void);
+
+#endif /* NETWORKING_H */
